+++ src/Neo.CLI/Settings.cs
@@ -13,8 +13,7 @@ using Microsoft.Extensions.Configuration;
 using Neo.Network.P2P;
 using Neo.Persistence;
 using System;
-using System.Linq;
-using System.Reflection;
+using System.Collections.Generic;
 using System.Threading;
 
 namespace Neo
@@ -26,7 +25,7 @@ namespace Neo
         public P2PSettings P2P { get; init; }
         public UnlockWalletSettings UnlockWallet { get; init; }
         public ContractsSettings Contracts { get; init; }
-        public PluginsSettings Plugins { get; init; }
+        public List<string>? Extras { get; init; }
 
         static Settings? s_default;
 
@@ -63,7 +62,7 @@ namespace Neo
             Storage = new(section.GetSection(nameof(Storage)));
             P2P = new(section.GetSection(nameof(P2P)));
             UnlockWallet = new(section.GetSection(nameof(UnlockWallet)));
-            Plugins = new(section.GetSection(nameof(Plugins)));
+            Extras = section.GetSection(nameof(Extras)).Get<List<string>>();
         }
 
         public Settings()
@@ -73,7 +72,7 @@ namespace Neo
             P2P = new P2PSettings();
             UnlockWallet = new UnlockWalletSettings();
             Contracts = new ContractsSettings();
-            Plugins = new PluginsSettings();
+            Extras = [];
         }
     }
 
@@ -162,26 +161,5 @@ namespace Neo
         }
 
         public ContractsSettings() { }
-    }
-
-    public class PluginsSettings
-    {
-        public Uri DownloadUrl { get; init; } = new("https://api.github.com/repos/neo-project/neo/releases");
-        public bool Prerelease { get; init; } = false;
-        public Version Version { get; init; } = Assembly.GetExecutingAssembly().GetName().Version!;
-
-        public PluginsSettings(IConfigurationSection section)
-        {
-            if (section.Exists())
-            {
-                DownloadUrl = section.GetValue(nameof(DownloadUrl), DownloadUrl)!;
-#if DEBUG
-                Prerelease = section.GetValue(nameof(Prerelease), Prerelease);
-                Version = section.GetValue(nameof(Version), Version)!;
-#endif
-            }
-        }
-
-        public PluginsSettings() { }
     }
 }
