+++ src/Neo.CLI/config.fs.mainnet.json
@@ -23,9 +23,15 @@
     "Contracts": {
       "NeoNameService": "0x7061fbd31562664b58f422c3dee4acfd70dba8af"
     },
-    "Plugins": {
-      "DownloadUrl": "https://api.github.com/repos/neo-project/neo/releases"
-    }
+    "Extras": [
+      "ApplicationLogs",
+      "DBFTPlugin",
+      "OracleService",
+      "RpcServer",
+      "StateService",
+      "StorageDumper",
+      "TokensTracker"
+    ]
   },
   "ProtocolConfiguration": {
     "Network": 91414437,
