+++ src/Neo.CLI/config.fs.testnet.json
@@ -23,9 +23,15 @@
     "Contracts": {
       "NeoNameService": "0xfb08ccf30ab534a871b7b092a49fe70c154ed678"
     },
-    "Plugins": {
-      "DownloadUrl": "https://api.github.com/repos/neo-project/neo/releases"
-    }
+    "Extras": [
+      "ApplicationLogs",
+      "DBFTPlugin",
+      "OracleService",
+      "RpcServer",
+      "StateService",
+      "StorageDumper",
+      "TokensTracker"
+    ]
   },
   "ProtocolConfiguration": {
     "Network": 91466898,
