+++ src/Neo.CLI/config.mainnet.json
@@ -23,9 +23,15 @@
     "Contracts": {
       "NeoNameService": "0x50ac1c37690cc2cfc594472833cf57505d5f46de"
     },
-    "Plugins": {
-      "DownloadUrl": "https://api.github.com/repos/neo-project/neo/releases"
-    }
+    "Extras": [
+      "ApplicationLogs",
+      "DBFTPlugin",
+      "OracleService",
+      "RpcServer",
+      "StateService",
+      "StorageDumper",
+      "TokensTracker"
+    ]
   },
   "ProtocolConfiguration": {
     "Network": 860833102,
