+++ src/Neo.CLI/configs/DBFTPlugin.json
