+++ src/Neo/NeoSystem.cs
@@ -107,8 +107,6 @@ namespace Neo
         {
             // Unify unhandled exceptions
             AppDomain.CurrentDomain.UnhandledException += CurrentDomain_UnhandledException;
-
-            Plugin.LoadPlugins();
         }
 
         /// <summary>
