+++ src/Plugins/ApplicationLogs/ApplicationLogs.csproj
@@ -16,12 +16,6 @@
   </ItemGroup>
 
   <ItemGroup>
-    <None Update="ApplicationLogs.json">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </None>
-  </ItemGroup>
-
-  <ItemGroup>
     <InternalsVisibleTo Include="$(PackageId).Tests" />
   </ItemGroup>
 </Project>
