+++ src/Plugins/DBFTPlugin/DBFTPlugin.csproj
@@ -11,10 +11,4 @@
     <ProjectReference Include="..\..\Neo.ConsoleService\Neo.ConsoleService.csproj" />
   </ItemGroup>
 
-  <ItemGroup>
-    <None Update="DBFTPlugin.json">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </None>
-  </ItemGroup>
-
 </Project>
