+++ src/Plugins/RpcServer/RpcServer.csproj
@@ -11,12 +11,6 @@
   </ItemGroup>
 
   <ItemGroup>
-    <None Update="RpcServer.json">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </None>
-  </ItemGroup>
-
-  <ItemGroup>
     <InternalsVisibleTo Include="Neo.Plugins.RpcServer.Tests" />
   </ItemGroup>
