+++ src/Plugins/StorageDumper/StorageDumper.csproj
@@ -11,11 +11,5 @@
   <ItemGroup>
     <ProjectReference Include="..\..\Neo.ConsoleService\Neo.ConsoleService.csproj" />
   </ItemGroup>
-  
-  <ItemGroup>
-    <None Update="StorageDumper.json">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </None>
-  </ItemGroup>
 
 </Project>
