+++ src/Plugins/TokensTracker/TokensTracker.cs
@@ -14,14 +14,13 @@ using Neo.IEventHandlers;
 using Neo.Ledger;
 using Neo.Network.P2P.Payloads;
 using Neo.Persistence;
-using Neo.Plugins.RpcServer;
 using Neo.Plugins.Trackers;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using static System.IO.Path;
 
-namespace Neo.Plugins
+namespace Neo.Plugins.TokensTracker
 {
     public class TokensTracker : Plugin, ICommittingHandler, ICommittedHandler
     {
@@ -78,7 +77,7 @@ namespace Neo.Plugins
             if (_enabledTrackers.Contains("NEP-17"))
                 trackers.Add(new Trackers.NEP_17.Nep17Tracker(_db, _maxResults, _shouldTrackHistory, neoSystem));
             foreach (TrackerBase tracker in trackers)
-                RpcServerPlugin.RegisterMethods(tracker, _network);
+                RpcServer.RpcServer.RegisterMethods(tracker, _network);
         }
 
         private void ResetBatch()
